tests: move stacks testing to its own application
Most of these checks can be performed on non-userspace
supporting platforms.
