+++ tests/kernel/mem_protect/userspace/src/test_syscall.h
@@ -7,8 +7,6 @@
 #ifndef USERSPACE_TEST_SYSCALL_H
 #define USERSPACE_TEST_SYSCALL_H
 
-__syscall void stack_info_get(char **start_addr, size_t *size);
-__syscall int check_perms(void *addr, size_t size, int write);
 __syscall void missing_syscall(void);
 __syscall void check_syscall_context(void);
