+++ tests/kernel/threads/thread_stack/prj.conf
@@ -1,0 +1,5 @@
+CONFIG_ZTEST=y
+CONFIG_INIT_STACKS=y
+CONFIG_THREAD_STACK_INFO=y
+CONFIG_APPLICATION_DEFINED_SYSCALL=y
+CONFIG_TEST_USERSPACE=y
